Separate v1,v2 payjoin-cli app mods
